Refactor cli module
