+++ src/mdns_beacon/cli/main.py
@@ -1,6 +1,6 @@
 """Console script for mdns-beacon."""
-from ipaddress import IPv4Address, IPv6Address, ip_address
-from typing import Any, AnyStr, Dict, Iterable, Optional, Union
+from ipaddress import IPv4Address, IPv6Address
+from typing import Any, Dict, Iterable, Union
 
 import click
 from rich.console import Console
@@ -8,6 +8,7 @@ from rich.table import Table
 from zeroconf import IPVersion, ServiceStateChange, Zeroconf
 
 from mdns_beacon import Beacon, BeaconListener, __version__
+from mdns_beacon.cli.types import IpAddressParamType
 
 console = Console()
 
@@ -21,27 +22,6 @@ TABLE_SERVICES_COLUMNS = [
     "Server",
     "TTL",
 ]
-
-
-class IpAddressParamType(click.ParamType):
-    """An IPv4Address or IPv6Address parsed via ipaddress.ip_address.
-
-    Example:
-        >>> ptype = IpAddressParamType()
-        >>> ptype.convert("127.0.0.1", None, None)
-        IPv4Address('127.0.0.1')
-    """
-
-    name = "ip_address"
-
-    def convert(
-        self, value: AnyStr, param: Optional[click.Parameter], ctx: Optional[click.Context]
-    ) -> Union[IPv4Address, IPv6Address]:
-        """Parse value into IPv4Address or IPv6Address."""
-        try:
-            return ip_address(value)
-        except ValueError:
-            self.fail(f"expected an IPv4 or IPv6 address, got {value!r}", param, ctx)
 
 
 @click.group()
